requirements.txt
mutagen
Pillow
python-dotenv
pydub==0.25.1
ttkthemes==3.2.2
pygame==2.5.2
mutagen==1.47.0
musicbrainzngs==0.7.1
simpleaudio==1.0.4
spotipy==2.23.0
